refactor: remove code duplication in validator classes
